feat: separated whatsapp button/interactive flow from common user message handling
