+++ functions/src/index.ts
@@ -14,6 +14,7 @@ export { healthcheckV2 } from "./definitions/healthcheck"
 export { onMessageWriteV2 } from "./definitions/eventHandlers/updateTypesense"
 export { onUserPublish } from "./definitions/eventHandlers/userHandlers"
 export { onCheckerPublish } from "./definitions/eventHandlers/checkerHandlerWhatsapp"
+export { onUserInteractivePublish } from "./definitions/eventHandlers/userHandlerWhatsappInteractive"
 export { telegramAuthHandler } from "./definitions/api/authentication"
 export { onCheckerPublishTelegram } from "./definitions/eventHandlers/checkerHandlerTelegram"
 export { apiHandler } from "./definitions/api/api"
